requirements.txt
# Meta API
requests>=2.28.2

# Google Cloud
google-cloud-bigquery>=3.9.0
google-cloud-secret-manager>=2.16.0
google-cloud-storage>=2.9.0

# Google Sheets
google-api-python-client>=2.79.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Data Processing
pandas>=1.5.3
numpy>=1.24.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.3.1